+++ nRF Toolbox/DFUOperations.m
@@ -421,7 +421,7 @@ NSDate *startTime, *finishTime;
 {
     finishTime = [NSDate date];
     self.uploadTimeInSeconds = [finishTime timeIntervalSinceDate:startTime];
-    NSLog(@"upload time in sec: %u",self.uploadTimeInSeconds);
+    NSLog(@"upload time in sec: %lu",(unsigned long)self.uploadTimeInSeconds);
 }
 
 #pragma mark - BLEOperations delegates
@@ -509,7 +509,7 @@ andControlPointCharacteristic:(CBCharacteristic *)dfuControlPointCharacteristic
 
 -(void)onFileOpened:(NSUInteger)fileSizeOfBin
 {
-    NSLog(@"onFileOpened file size: %d",fileSizeOfBin);
+    NSLog(@"onFileOpened file size: %lu",(unsigned long)fileSizeOfBin);
     binFileSize += fileSizeOfBin;
 }
